link with ocaml-compiler-libs to use the OCaml parser
(the parser is not really plumbed into the project yet)

We use ocaml-migrate-parsetree which provides a versioned copy of the
OCaml AST (we use Ast_408, corresponding to the definition of OCaml
4.08), so that the type definitions do not change if we build our
project under a different OCaml version.

(The current-version AST is converted automatically by
ocaml-migrate-parsetree into the fixed-version AST we are using, and
this only fails if the input uses un-translatable AST constructs,
which should not be the case in this project, where our OCaml code
inputs are mostly simple pattern-matching definitions).