a more semantic representation of target piops
With this commit, target symbolic constraints are represented directly
by their "domain", that is the set of all possible concrete values.

The concrete value of an accessor may be a value of the form (Int n)
or a value of the form (Tag n), so in the general case a symbolic
value may belong to a set of possible integers *and* a set of possible
tags: a pair of sets.

Intuitively, one can efficiently represent those sets because they are
created by simple operations (equalities and comparisons
with integers), negation and intersection, so they remain unions of
intervals that can be implemented efficiently. (We use an external
library called Diet to represent arbitrary unions of integer
intervals.)

It should now be very simple to implement logical simplifications on
those domains (intersection, equality or inclusion checking, etc.), by
directly using the operations on the underlying sets.