we can get rid of \covers !
In our implementation, target fallback nodes carry a domain in
addition to their decision tree. This is what makes \covers
interesting: the union the domain of all cihld need not cover
everything.

In the workshop presentation, we do *not* carry a domain in target
decision trees (to make them more similar to source decision trees),
and it turns out that this work just as well (the trees are less
precise, but the missing information can be found in the
input domain).

With this presentation, all nodes naturally cover the whole domain, so
the extra "covers D S" condition is not needed anymore.